refactor: Update Modal import path and improve ExperienceSummary component structure
